Refactor app/main (former pages/document)
