+++ public/files/js/pages/subcorpForm.ts
@@ -30,7 +30,7 @@ import {init as subcorpViewsInit} from 'views/subcorp/forms';
 import {SubcorpWithinFormStore, SubcorpFormStore} from '../stores/subcorp/form';
 import liveAttributes from 'plugins/liveAttributes/init';
 import subcMixer from 'plugins/subcmixer/init';
-import {UserSettings} from '../userSettings';
+import {UserSettings} from '../app/userSettings';
 import {TextTypesStore} from '../stores/textTypes/attrValues';
 import {init as ttViewsInit} from 'views/textTypes';
 import * as corplistComponent from 'plugins/corparch/init'
