+++ public/files/js/plugins/ucnkSyntaxViewer/ucnkTreeView.ts
@@ -607,7 +607,7 @@ export function generate(data:Array<SourceData.Data>, zone:string, tree:string, 
         addGlobalKeyEventHandler:(fn:(evt:Event)=>void):void => {},
         removeGlobalKeyEventHandler:(fn:(evt:Event)=>void):void => {},
         cloneState:<T extends {[key:string]:any}>(obj:T):T => obj,
-        delayHandler:(immediateFn:()=>void, actualFn:()=>void, delay:number) => undefined,
+        doThingsWithDelay:(immediateFn:()=>void, actualFn:()=>void, delay:number) => undefined,
         getHelpLink:(ident:string) => ''
     };
     const gen = new TreeGenerator(options, helpers);
