+++ public/files/js/stores/concordance/lineSelection.ts
@@ -126,7 +126,7 @@ export class LineSelectionStore extends SimplePageStore {
                             this._isBusy = false;
                             this.concLineStore.notifyChangeListeners();
                             this.notifyChangeListeners();
-                            this.layoutModel.history.replaceState('view', args);
+                            this.layoutModel.getHistory().replaceState('view', args);
                         },
                         (err) => {
                             this._isBusy = false;
@@ -151,7 +151,7 @@ export class LineSelectionStore extends SimplePageStore {
                             this._isBusy = false;
                             this.notifyChangeListeners();
                             this.concLineStore.notifyChangeListeners();
-                            this.layoutModel.history.replaceState('view', args);
+                            this.layoutModel.getHistory().replaceState('view', args);
                         },
                         (err) => {
                             this._isBusy = false;
@@ -171,7 +171,7 @@ export class LineSelectionStore extends SimplePageStore {
                             this._isBusy = false;
                             this.concLineStore.notifyChangeListeners();
                             this.notifyChangeListeners();
-                            this.layoutModel.history.replaceState('view', args);
+                            this.layoutModel.getHistory().replaceState('view', args);
                         },
                         (err) => {
                             this._isBusy = false;
@@ -188,7 +188,7 @@ export class LineSelectionStore extends SimplePageStore {
                             this._isBusy = false;
                             this.concLineStore.notifyChangeListeners();
                             this.notifyChangeListeners();
-                            this.layoutModel.history.replaceState('view', args);
+                            this.layoutModel.getHistory().replaceState('view', args);
                         },
                         (err) => {
                             this._isBusy = false;
