+++ public/files/js/stores/concordance/lines.ts
@@ -465,7 +465,7 @@ export class ConcLineStore extends SimplePageStore {
     private pushHistoryState(pageNum:number):void {
         const args = this.layoutModel.getConcArgs();
         args.set('fromp', pageNum);
-        this.layoutModel.history.pushState(
+        this.layoutModel.getHistory().pushState(
             'view', args, { pagination: true, pageNum: pageNum });
     }
