+++ public/files/js/util.ts
@@ -250,83 +250,6 @@ export class MultiDict implements Kontext.IMultiDict {
     }
 }
 
-/**
- * NullHistory is a fallback object to be used
- * in browsers where HTML5 history is not available.
- * (but the truth is, it won't help much anyway
- * as many different issues remain unsolved in
- * case of outdated browsers)
- */
-export class NullHistory implements Kontext.IHistory {
-    replaceState(action:string, args:Kontext.IMultiDict, stateData?:any, title?:string):void {}
-    pushState(action:string, args:Kontext.IMultiDict, stateData?:any, title?:string):void {}
-    setOnPopState(fn:(event:{state: any})=>void):void {}
-}
-
-/**
- * A simple wrapper around window.history object
- * with more convenient API.
- */
-export class History implements Kontext.IHistory {
-
-    private h:Kontext.IURLHandler;
-
-    constructor(urlHandler:Kontext.IURLHandler) {
-        this.h = urlHandler;
-    }
-
-    /**
-     * Replace the current state with the one specified by passed arguments.
-     *
-     * @param action action name (e.g. 'first_form', 'subcorpus/subcorp_list')
-     * @param args a multi-dict instance containing URL arguments to be used
-     * @param stateData (just like in window.history.replaceState)
-     * @param title (just like in window.history.replaceState), default is window.document.title
-     */
-    replaceState(action:string, args:Kontext.IMultiDict, stateData?:any, title?:string):void {
-        if (/^https?:\/\//.exec(action)) {
-            throw new Error('Invalid action specifier (cannot use URL here)');
-        }
-        window.history.replaceState(
-            stateData || {},
-            title || window.document.title,
-            `${this.h.createActionUrl(action)}?${this.h.encodeURLParameters(args)}`
-        );
-    }
-
-    /**
-     * Push a new state
-     *
-     * @param action action name (e.g. 'first_form', 'subcorpus/subcorp_list')
-     * @param args a multi-dict instance containing URL arguments to be used
-     * @param stateData (just like in window.history.replaceState)
-     * @param title (just like in window.history.replaceState), default is window.document.title
-     */
-    pushState(action:string, args:Kontext.IMultiDict, stateData?:any, title?:string):void {
-        if (/^https?:\/\//.exec(action)) {
-            throw new Error('Invalid action specifier (cannot use URL here)');
-        }
-        window.history.pushState(
-            stateData || {},
-            title || window.document.title,
-            `${this.h.createActionUrl(action)}?${this.h.encodeURLParameters(args)}`
-        );
-    }
-
-    setOnPopState(fn:(event:{state: any})=>void):void {
-        window.onpopstate = fn;
-    }
-}
-
-export function createHistory(urlHandler:Kontext.IURLHandler):Kontext.IHistory {
-    if (window.history && 'pushState' in window.history) {
-        return new History(urlHandler);
-
-    } else {
-        return new NullHistory();
-    }
-}
-
 
 export function uid():string {
     return `${new Date().getTime()}${Math.random()}`.substr(2);
